+++ changelog.md
@@ -35,6 +35,8 @@ Codion Change Log
 - EntityPanel.USE_KEYBOARD_NAVIGATION renamed KEYBOARD_NAVIGATION, Config.useKeyboardNavigation renamed keyboardNavigation.
 - DefaultFrameworkIcons bug fixed, logo color now updated along with other icons, related refactoring.
 - FrameworkIcons.logo(size) removed.
+- DefaultFrameworkIcons no longer contains a standard sized LOGO, unused.
+- FrameworkIcons.refreshRequired() removed.
 ### is.codion.tools.generator.model
 - DomainGeneratorModel.DEFAULT_DOMAIN_PACKAGE and DEFAULT_SOURCE_DIRECTORY renamed DOMAIN_PACKAGE and SOURCE_DIRECTORY.